chore: refactor components placement and naming
