+++ packages/apps/spaces/.env.development
@@ -28,6 +28,8 @@ NEXT_PUBLIC_WEBSOCKET_PATH=ws://127.0.0.1:8006/
 NEXT_PUBLIC_TURN_SERVER=turn:127.0.0.1:3478?transport=tcp
 NEXT_PUBLIC_TURN_USER=foo
 NEXT_PUBLIC_TURN_CREDENTIAL=1234
+NEXT_PUBLIC_SSR_PUBLIC_PATH=http://localhost:3001
+NEXT_PUBLIC_CUSTOMER_OS_API_PATH=http://127.0.0.1:10000
 
 WEB_CHAT_API_KEY=nobody-will-guess-this-api-key
 WEB_CHAT_HTTP_PATH=http://127.0.0.1:8013/api/v1
