+++ packages/apps/spaces/app/(routes)/organizations/(routes)/[id]/components/FormSocialInput/SocialIcons.tsx
@@ -1,0 +1,16 @@
+import { Icons } from '@ui/media/Icon';
+
+import { isKnownUrl } from './util';
+
+export const SocialIcon = ({
+  children,
+  url,
+}: React.PropsWithChildren<{ url: string }>) => {
+  const knownUrl = isKnownUrl(url);
+
+  if (knownUrl === 'twitter')
+    return <Icons.Twitter viewBox='0 0 32 32' strokeWidth='0' />;
+  if (knownUrl === 'linkedin')
+    return <Icons.Linkedin viewBox='0 0 32 32' strokeWidth='0' />;
+  return <>{children}</>;
+};
