+++ packages/apps/spaces/app/(routes)/organizations/(routes)/[id]/components/OrganizationAboutTab/index.ts
@@ -1,0 +1,1 @@
+export * from './OrganizationAboutTab';
