+++ packages/apps/spaces/app/(routes)/organizations/(routes)/[id]/components/OrganizationAccountTab/OrganizationAccountTab.tsx
@@ -1,0 +1,9 @@
+import { TabPanel } from '@ui/disclosure/Tabs';
+
+export const OrganizationAccountTab = () => {
+  return (
+    <TabPanel>
+      <p>Account</p>
+    </TabPanel>
+  );
+};
