--- packages/apps/spaces/app/(routes)/organizations/(routes)/[id]/components/OrganizationInfo/index.ts
@@ -1,4 +1,0 @@
-export * from './OrganizationInfo';
-export * from './OrganizationHeader';
-export * from './OrganizationTabs';
-export * from './OrganizationDetails';
