+++ packages/apps/spaces/app/(routes)/organizations/(routes)/[id]/components/OrganizationPeopleTab/OrganizationPeopleTab.tsx
@@ -1,0 +1,9 @@
+import { TabPanel } from '@ui/disclosure/Tabs';
+
+export const OrganizationPeopleTab = () => {
+  return (
+    <TabPanel>
+      <p>People</p>
+    </TabPanel>
+  );
+};
