+++ packages/apps/spaces/app/(routes)/organizations/(routes)/[id]/components/OrganizationPeopleTab/index.ts
@@ -1,0 +1,1 @@
+export * from './OrganizationPeopleTab';
