+++ packages/apps/spaces/app/(routes)/organizations/(routes)/[id]/components/OrganizationSuccessTab/OrganizationSuccessTab.tsx
@@ -1,0 +1,9 @@
+import { TabPanel } from '@ui/disclosure/Tabs';
+
+export const OrganizationSuccessTab = () => {
+  return (
+    <TabPanel>
+      <p>Success</p>
+    </TabPanel>
+  );
+};
