+++ packages/apps/spaces/app/(routes)/organizations/(routes)/[id]/components/OrganizationSuccessTab/index.ts
@@ -1,0 +1,1 @@
+export * from './OrganizationSuccessTab';
