+++ packages/apps/spaces/app/(routes)/organizations/(routes)/[id]/components/OrganizationTabs/OrganizationTabsHeader.tsx
@@ -3,7 +3,7 @@
 import { Image } from '@ui/media/Image';
 import { Flex } from '@ui/layout/Flex';
 
-export const OrganizationHeader = ({
+export const OrganizationTabsHeader = ({
   children,
 }: {
   children?: React.ReactNode;
