+++ packages/apps/spaces/app/(routes)/organizations/(routes)/[id]/components/OrganizationUpNextTab/OrganizationUpNextTab.tsx
@@ -1,0 +1,9 @@
+import { TabPanel } from '@ui/disclosure/Tabs';
+
+export const OrganizationUpNextTab = () => {
+  return (
+    <TabPanel>
+      <p>Up Next</p>
+    </TabPanel>
+  );
+};
