+++ packages/apps/spaces/app/(routes)/organizations/(routes)/[id]/components/OrganizationUpNextTab/index.ts
@@ -1,0 +1,1 @@
+export * from './OrganizationUpNextTab';
